refactor(can-tx): raw loopback via CanDb_PackSignal + sweep signedness
TX path now operates on the raw bus u32 instead of physical s32; this avoids

a needless (value - offset) / factor round-trip inside the per-signal pack.

Code changes:

- CanTx_EncodeSignal(can_id, sig_id, u32 raw): internal CanDb_PackSignal()

- CanTx_RebuildFromSignals: Signal_Get() is u32 raw, pack via CanDb_PackSignal

- CanDb_EncodeAndPack docstring: clarified as physical-only path;

  raw-loopback should use CanDb_PackSignal / CanDb_BitEncode

- can_if.c: drop redundant (int32_t) cast on u32 s_bus_off_count

- mod_can_demo.c demo #5: u32 v = (sweep * 7u) % 512u

Verified remaining signedness usage in app/ as intentional:

- types.h: sN/intN typedef aliases

- result.h: c02b2_result_t (negative error codes)

- rti_defer.c: (int32_t)(now - deadline) for wrap-safe delta

- can_db_codec.h/.c: DBC sign-extend + bit math + (rare) physical API

+++ app/drv_api/can/can_if.c
@@ -331,7 +331,7 @@ static void prv_flexcan_err_cb(u8 instance,
             s_bus_off_count[ch]++;
             (void)Signal_Set(SIG_CAN_BUS_OFF, 1);
             (void)Signal_Set(SIG_CAN_BUS_OFF_COUNT,
-                             (int32_t)s_bus_off_count[ch]);
+                             s_bus_off_count[ch]);
             LOG_E("CAN%u BUS_OFF enter (tx_err=%u rx_err=%u, total=%u)",
                   (unsigned)instance,
                   (unsigned)tx_err, (unsigned)rx_err,
